#130: Make the JS render the followups only on user request.
Rendering followups server-side by default makes the request hangs for a
lot of time (more than 2 seconds), this can be a hassle when a user may
want to check the list without looking at anything but the tasks.
This commit makes the client render the followups when a user wants to
display the followups by expanding the details of a task.