Rewrite updateFollowups function
The previous version of updateFollowups was only doing an append of the
latest followup on the current followups list. This behavior leads to
comment duplication in the DOM and weird behaviors.
This is a full rewrite of the function that will just clear the
followups list and repopulate it with the proper list of followups.