Views: add a reject_task view
This commit adds a way to reject a task.
Before this commit, the only way to discard a commit was to mark it as
done. The problem is that some tasks can't be done for any reasons, or
need to be rewritten, or are obsolete, etc.
This commit allows a user to mark a task as rejected, add a followup to
describe how and why the task is rejected.