WIP: Add a dedicated view for a single task
This commit adds a dedicated view for a single task, this feature aims
to be fully compatible with the current way of dealing with tasks.

This shall allow a user to manage a single task without having to deal
with the current table view. This will also allow a user to handle
subtasks by commenting them, updating their descriptions, etc.

This commit also contains a started rework of how javascript functions
are handled and a refactor of them: since the dedicated view and the
table page of a task have a lot of functions in common, it's better to
group these functions inside a common file instead of rewriting the same
code over and over.