Fix Task update behavior
When a task was updated by the user, the followup type could be a
STATE_CHANGE or a COMMENT, however during a task edition the followup
type should be STATE_CHANGE or MODIFICATION since it is not possible in
the front to add a followup while editing the task.

This commit fixes this behavior.