#148: Hide the creation, resolution and due date columns
This commit adds the `hidden` class for the following columns :
	* Creation date
	* Resolution date
	* Due date

The reason being that these columns may display useful informations but
most of the time they hurt the readability of the task details by
distorting the table more than needed.

This commit doesn't removes these columns as we will still use them in
the future by allowing the users to choose between not displaying them
and display them.

The display/hide feature will come in future commit.