#150: Add "hidden" class to newly created tasks
The previous commit (see `ebcc569`) introduced a new way to visualize
tasks, by default the dates related to a task (creation, completion and
deadline) are now hidden to the user.

However this change wasn't properly backported into the JS code that
handles the creation of a new task and its addition to the DOM.
This commit fixes that.