chore: refactor routes validators
This commit refactors the way validators work: previously the validators
would modify the attribute ``task`` of the incoming request to pass the
``Task`` object from the validators to the decorated functions.

However, this behavior hides the retrieval of the object and the used
validators are doing more than validation (since they're also
responsible for retrieval).

This commit changes this by:
	* Using the ``exists`` statement instead of relying on
	  ``DoesNotExists`` exception to check if the task exists in
	  database.
	* Relocating the responsability of object retrieval to the view
	  and not the validators themselves