tests: rewrite deletion test case with factories
This commit edits the current test case for todo list deletion by adding
the usage of factories instead of forging the data "by hand" in the
setup method.

This commit also adds the usage of ``HTTPStatus`` instead of using bare
ints in order to improve clarity on what is to be expected out of the
API calls.