refactor: unify how tasks are retrieved for single tasks
This commit moves the task retrieval in a single function instead of
having each route replicate the same behavior.
It also fixes a pattern that should've been corrected a long time ago:
we do not want to rely on exceptions to check whether an object exists
in the database or not.

The new retrieval function is here to:
  * check if the provided ids are not null
  * check if the task actually exists in database

If one of the previous conditions aren't matched then the return value
is ``None`` and each view will return a 404.