fix: Make id and task_no optional in TaskSchema
``id`` is an autogenerated and autoincremented field defined at the
first instantiation of an instance.

``task_no`` is a computed value that is incremented based on the number
of tasks in a given list.

These two fields are not needed at the creation of task, however the
previous implementation of the TaskSchema did not had these fields
marked as optional, making creation tests fail.